% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.csb_ewma}
\alias{plot.csb_ewma}
\title{Plot CSB-EWMA Control Chart}
\usage{
\method{plot}{csb_ewma}(x, title = "CSB-EWMA Control Chart", show_signal = TRUE, ...)
}
\arguments{
\item{x}{csb_ewma object from csb_ewma() function}

\item{title}{Plot title (default = "CSB-EWMA Control Chart")}

\item{show_signal}{Whether to highlight signal point (default = TRUE)}

\item{...}{Additional arguments passed to ggplot}
}
\value{
A ggplot object (invisibly) and displays the plot
}
\description{
Creates a professional control chart showing EWMA statistic and limits.
}
\examples{
# See csb_ewma() for examples
}
