% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{print.csb_ewma}
\alias{print.csb_ewma}
\title{Prints a formatted summary of CSB-EWMA chart results.}
\usage{
\method{print}{csb_ewma}(x, ...)
}
\arguments{
\item{x}{csb_ewma object from csb_ewma() or run_csb_ewma() function}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the object
}
\description{
Prints a formatted summary of CSB-EWMA chart results.
}
