% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rlaplace}
\alias{rlaplace}
\title{Generate Laplace (Double Exponential) Random Variables}
\usage{
rlaplace(n, location = 0, scale = 1)
}
\arguments{
\item{n}{Number of observations to generate}

\item{location}{Location parameter (median) of the distribution (default = 0)}

\item{scale}{Scale parameter (spread) of the distribution (default = 1)}
}
\value{
A numeric vector of length n containing Laplace random variables
}
\description{
Generates random numbers from the Laplace distribution using inverse transform sampling.
}
\examples{
\dontrun{
x <- rlaplace(100, location = 0, scale = 1)
}
----------------------------------------------------------------------------
}
